agent/requirements.txt
psutil>=5.9.0
requests>=2.28.0
pystray>=0.19.0
Pillow>=9.0.0
